Fixes integrador2
acomodamos la clave dni como PK, por lo que cambiamos metodos como el insertar estudiante y carrera para que funcione, push ivan y mai

+++ Integrador2_grupo8/inte/src/main/java/org/integrador/repository/EstudianteDeCarreraRepository.java
@@ -28,7 +28,7 @@ public class EstudianteDeCarreraRepository implements Repository<EstudianteDeCar
         }
 
         @Override
-        public EstudianteDeCarrera findById(long id) {
+        public EstudianteDeCarrera findById(int id) {
             EstudianteDeCarrera e = em.find(EstudianteDeCarrera.class, id);
             if (e != null) {
                 System.out.println("Encontrado: " + e);
