+++ Integrador2_grupo8/inte/src/main/java/org/integrador/repository/Repository.java
@@ -9,6 +9,6 @@ public interface Repository <T>{
 
 
     void delete(int id);
-    T findById(long id);
+    T findById(int id);
     List<T> findAll();
 }
